- Addded functionality for rendering text. To display text create a GUIText object. Then run textMaster.loadText(GUIText Object); Render it in the render function in Renderer.java class using textMaster.render();
BRIEF SUMMARY
The way this works is by creating a .fnt file that outlines the spacing and coordinates for all the characters in the font file.
Then load the png file with all the characters into a texture using the PNG Decoder.
The .fnt file is loaded into the system and the system then uses this to pull parts out of the PNG file.
This is then shown to screen.

You will need to refresh the Gradle as there are new depositories in order for this to work properly..

PLEASE NOTE:
I've still got to clean all this up a bit but it works well!