Create a primitive particle system.
The particle system renders 100 pixel particles with size 4. The particles are
moved according to a force field and their colour grows colder over time. Particles
have a 45% chance of coming alive each frame after they have died.

The coordinate space of the particles is relative to the particle system (i.e. the
origin for a particle is the position of the particle system in the world).