Render using nested coordinate spaces.
All of the coordinate spaces for any Node, or GameItem are the same.

A Node can have multiple child Nodes, and changes to a parent's position and rotation will also affect its children.

The Camera is also a Node, which means that if you wanted to say, position something so it was always in view of the camera, you could attach it as a child of the camera and it would show up in a fixed position.

A Node with multiple children replaces GameItemGroup.