Rings are now placed around the map correctly. This is done from loading in x pools of data corresponding to x amount of rings from a txt file. The program then selects an option randomly from each pool for each ring. Therefore also added in a file reader class and a data file.
Added in boost functionality. Needs some restrictions as player can infinitely boost but the core concept is there including sound and increased arrow speed.

Updated WAV Player to stop sound as soon as the playing boolean is set to false for that specific track. This is done via the stopSound() method.